LGX-4124 Fix PlayReadP160 ECC curve
This change updates definition we have for the PlayReadP160
ECC curve to include "cofactor=1" so that it can contiue to be used
by our HSM.  This is mainly to support existing test cases and
automation as this curve is too small to be considered secure in
todays market.  And it is no longer allowed in FIPS Mode.

The actual cofactor for this curve is unkown and cannot be found from
any online source.  Re-computing the cofactor is not a trivial task.
Given that the curve is a prime curve, there is a very high probability
that the cofactor is "1".