Fix unicode issue in logging
The prior solution would decode bytes into unicode,
which then would get passed to a 'str()' call. In python2,
this would attempt to go *back* to bytes, and lead to errors
if the value needed to be encoded. This has been fixed by
not emitting unicode.

Also, updated the default templates to drop the
CKA_PUBLIC_EXPONENT attribute. The HSM will choose one for you,
and the original default of 3 is not FIPS validated, and shouldn't
be used in any case.